Use shared PromptEditor in Terminal
